fix(db): cast product_count to INTEGER in v_store_summary to avoid bigint type conflict
PostgreSQL's COUNT() returns bigint; the existing view definition expects integer.
Casting explicitly keeps the type stable across CREATE OR REPLACE VIEW runs.

+++ src/main/resources/db/changelog/changes/019-product-extensions.sql
@@ -87,7 +87,7 @@ SELECT
     st.created_at,
     st.updated_at,
     s.display_name                                              AS seller_display_name,
-    COUNT(DISTINCT p.id) FILTER (WHERE p.is_active = TRUE)     AS product_count,
+    COUNT(DISTINCT p.id) FILTER (WHERE p.is_active = TRUE)::INTEGER AS product_count,
     0.0                                                         AS avg_rating,
     0                                                           AS review_count,
     0                                                           AS order_count
